Split environments.py into multiple files
Having all environments in one module would result in a massive file.

Split into a submodule where each environment can get their own file.
